+++ lib/invoicexpress/models.rb
@@ -25,5 +25,4 @@ require 'invoicexpress/models/top_debtor'
 require 'invoicexpress/models/quarterly_result'
 require 'invoicexpress/models/schedule'
 require 'invoicexpress/models/purchase_order'
-
-
+require 'invoicexpress/models/supplier'