+++ lib/invoicexpress/models/invoice.rb
@@ -134,6 +134,19 @@ module Invoicexpress
       tag 'debit_note'
     end
 
+    #added for app.invoicexpress api version
+    class CoreInvoiceReceipt < BaseModel
+      include BaseInvoice
+      tag 'invoice_receipt'
+    end
+
+    #added for app.invoicexpress api version
+    class InvoiceReceipt < BaseModel
+      include BaseInvoice
+      include ExtraInvoice
+      tag 'invoice_receipt'
+    end
+
     class SimplifiedInvoice < BaseModel
       include BaseInvoice
       include ExtraInvoice
@@ -182,6 +195,15 @@ module Invoicexpress
       element :total_pages, Integer
       element :total_entries, Integer
       element :per_page, Integer
+    end
+    
+    #added for app.invoicexpress api version
+    class InvoiceReceipts < BaseModel
+      include HappyMapper
+
+      tag 'invoice_receipts'
+      has_one :results, InvoiceResult
+      has_many :invoice_receipts, CashInvoice
     end
 
     class CashInvoices < BaseModel